Move all files to nylas/ folder.
Nylas files pollute global ruby *file* namespace.
For example if i have gem called 'account' or 'contact' it could be a
problem.
