--- lib/label.rb
@@ -1,4 +1,0 @@
-require 'folder'
-
-module Nylas
-end
