+++ lib/nylas.rb
@@ -7,22 +7,22 @@ require 'em-http'
 require 'ostruct'
 require 'active_support/core_ext/hash'
 
-require 'account'
-require 'api_account'
-require_relative 'thread'
-require 'calendar'
-require 'account'
-require 'message'
-require 'expanded_message'
-require 'draft'
-require 'contact'
-require 'file'
-require 'calendar'
-require 'event'
-require 'folder'
-require 'restful_model'
-require 'restful_model_collection'
-require 'version'
+require 'nylas/account'
+require 'nylas/api_account'
+require 'nylas/thread'
+require 'nylas/calendar'
+require 'nylas/account'
+require 'nylas/message'
+require 'nylas/expanded_message'
+require 'nylas/draft'
+require 'nylas/contact'
+require 'nylas/file'
+require 'nylas/calendar'
+require 'nylas/event'
+require 'nylas/folder'
+require 'nylas/restful_model'
+require 'nylas/restful_model_collection'
+require 'nylas/version'
 
 module Nylas
   Error = Class.new(::StandardError)
