+++ lib/nylas/expanded_message.rb
