+++ lib/nylas/label.rb
@@ -1,0 +1,4 @@
+require 'nylas/folder'
+
+module Nylas
+end
