+++ lib/nylas/mixins.rb
