+++ spec/event_spec.rb
@@ -1,4 +1,4 @@
-require 'event'
+require 'nylas/event'
 
 describe Nylas::Event do
   before (:each) do
